Move common client logic into Client::Implementation
